# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервис для отслеживания статистики игроков в дуэлях.",
        "title": "Player Stats API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/players/{name}": {
            "patch": {
                "security": [
                    {
                        "secret": []
                    }
                ],
                "description": "Частичное обновление данных игрока (убийства, смерти, матчи). Можно передать только те поля, которые нужно изменить.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "players"
                ],
                "summary": "Обновить статистику игрока",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя игрока",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления (JSON)",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.UpdatePlayer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное обновление",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный JSON или параметры",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Игрок не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации (например, отрицательные значения)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.UpdatePlayer": {
            "type": "object",
            "properties": {
                "deaths": {
                    "type": "integer"
                },
                "kills": {
                    "type": "integer"
                },
                "matches": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "secret": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}